@@ -5,7 +5,7 @@ description = "Default template for PDM package"
 authors = [{ name = "svenrdz", email = "svenbmr@gmail.com" }]
 dependencies = ["prefect[docker]>=3.0.3"]
 readme = "README.md"
-requires-python = ">= 3.9"
+requires-python = ">= 3.10"
 license = { text = "MIT" }
 
 [project.entry-points."prefect.collections"]
